hacker_news/requirements.txt
langchain
langchain_openai
python-dotenv
streamlit
PyPDF2